remove circular dependency in libs
