+++ apps/hash-api/src/graph/ontology/primitive/data-type.ts
@@ -9,13 +9,13 @@ import {
   ModifyRelationshipOperation,
   OntologyTemporalMetadata,
 } from "@local/hash-graph-client";
-import { ConstructDataTypeParams } from "@local/hash-isomorphic-utils/data-types";
 import {
   currentTimeInstantTemporalAxes,
   zeroedGraphResolveDepths,
 } from "@local/hash-isomorphic-utils/graph-queries";
 import { generateTypeId } from "@local/hash-isomorphic-utils/ontology-types";
 import {
+  ConstructDataTypeParams,
   DataTypeAuthorizationRelationship,
   DataTypeMetadata,
   DataTypeRelationAndSubject,
