+++ libs/@local/hash-subgraph/package.json
@@ -33,7 +33,6 @@
     "@blockprotocol/type-system": "0.1.1",
     "@local/advanced-types": "0.0.0-private",
     "@local/hash-graph-client": "0.0.0-private",
-    "@local/hash-isomorphic-utils": "0.0.0-private",
     "uuid": "9.0.1"
   },
   "devDependencies": {
